+++ misc/changelog.md
@@ -15,6 +15,7 @@
 - Update ExifTool to `13.25`.
 - Update ImageMagick to `7.1.1-46`.
 - Adjust update checker to Flatpak.
+- Notify user when Flatpak permissions are insufficient.
 
 ### Fixed